tools/deal-pack-cli/requirements.txt
openpyxl>=3.1,<4.0
pdfplumber>=0.11,<1.0
Pillow>=10.0
pytest>=8.0
pytest-cov>=5.0
